+++ moauthd/client.c
@@ -224,6 +224,8 @@ moauthdRunClient(
 
     if ((authorization = httpGetField(client->http, HTTP_FIELD_AUTHORIZATION)) != NULL && *authorization)
     {
+      moauthdLogc(client, MOAUTHD_LOGLEVEL_DEBUG, "Authorization: %s", authorization);
+
       if (!strncmp(authorization, "Basic ", 6))
       {
         // Basic authentication...
@@ -281,7 +283,9 @@ moauthdRunClient(
         // Bearer (OAuth) token...
         moauthd_token_t *token;		// Access token
 
-        for (authorization += 7; *authorization && isspace(*authorization & 255); authorization ++);
+        authorization += 7;
+        while (*authorization && isspace(*authorization & 255))
+          authorization ++;
 
         if ((token = moauthdFindToken(client->server, authorization)) != NULL)
         {
@@ -452,8 +456,6 @@ do_authorize(moauthd_client_t *client)	// I - Client object
   const char	*prefix;		// Prefix string
 
 
-  moauthdLogc(client, MOAUTHD_LOGLEVEL_DEBUG, "state=%s", httpStateString(httpGetState(client->http)));
-
   switch (client->request_method)
   {
     case HTTP_STATE_HEAD :
@@ -461,8 +463,6 @@ do_authorize(moauthd_client_t *client)	// I - Client object
 
     case HTTP_STATE_GET :
         // Get form variable on the request line...
-        moauthdLogc(client, MOAUTHD_LOGLEVEL_DEBUG, "query_string=\"%s\"", client->query_string);
-
         num_vars      = cupsFormDecode(client->query_string, &vars);
         client_id     = cupsGetOption("client_id", num_vars, vars);
         redirect_uri  = cupsGetOption("redirect_uri", num_vars, vars);
@@ -471,13 +471,6 @@ do_authorize(moauthd_client_t *client)	// I - Client object
         state         = cupsGetOption("state", num_vars, vars);
         challenge     = cupsGetOption("code_challenge", num_vars, vars);
         method        = cupsGetOption("code_challenge_method", num_vars, vars);
-
-        {
-          size_t i;
-          moauthdLogc(client, MOAUTHD_LOGLEVEL_DEBUG, "num_vars=%u", (unsigned)num_vars);
-          for (i = 0; i < num_vars; i ++)
-            moauthdLogc(client, MOAUTHD_LOGLEVEL_DEBUG, "vars[%u].name=\"%s\", .value=\"%s\"", (unsigned)i, vars[i].name, vars[i].value);
-        }
 
         if (!client_id || !response_type || strcmp(response_type, "code") || (method && strcmp(method, "S256")))
         {
@@ -574,8 +567,6 @@ do_authorize(moauthd_client_t *client)	// I - Client object
         if ((data = _moauthCopyMessageBody(client->http)) == NULL)
           return (moauthdRespondClient(client, HTTP_STATUS_BAD_REQUEST, NULL, NULL, 0, 0));
 
-        moauthdLogc(client, MOAUTHD_LOGLEVEL_DEBUG, "body=\"%s\"", data);
-
         num_vars      = cupsFormDecode(data, &vars);
         client_id     = cupsGetOption("client_id", num_vars, vars);
         redirect_uri  = cupsGetOption("redirect_uri", num_vars, vars);
@@ -585,13 +576,6 @@ do_authorize(moauthd_client_t *client)	// I - Client object
         username      = cupsGetOption("username", num_vars, vars);
         password      = cupsGetOption("password", num_vars, vars);
         challenge     = cupsGetOption("code_challenge", num_vars, vars);
-
-        {
-          size_t i;
-          moauthdLogc(client, MOAUTHD_LOGLEVEL_DEBUG, "num_vars=%u", (unsigned)num_vars);
-          for (i = 0; i < num_vars; i ++)
-            moauthdLogc(client, MOAUTHD_LOGLEVEL_DEBUG, "vars[%u].name=\"%s\", .value=\"%s\"", (unsigned)i, vars[i].name, vars[i].value);
-        }
 
 	free(data);